chore: replace apiDefineComponet
