+++ test-dts/defineComponent.test-d.tsx
@@ -743,31 +743,3 @@ describe('componentOptions setup should be `SetupContext`', () => {
     ctx: SetupContext
   ) => any)
 })
-
-const Wprops = defineComponent({
-  props: {
-    x1: {
-      type: String,
-      required: true
-    },
-    x2: Number
-  }
-})
-
-const Xprops = newDefineComponent({
-  props: {
-    x1: {
-      type: String,
-      required: true
-    },
-    x2: Number
-  }
-})
-
-// @ts-expect-error
-expectType<JSX.Element>(<Xprops />)
-
-// @ts-expect-error
-expectError(<Xprops />)
-// @ts-expect-error
-expectError(<Wprops />)
